+++ src/main/java/com/botmaker/sdk/internal/plugin/editors/GeometryEditors.java
@@ -3,6 +3,7 @@ package com.botmaker.sdk.internal.plugin.editors;
 import com.botmaker.plugin.api.ValueContext;
 import com.botmaker.plugin.toolkit.Modals;
 import com.botmaker.plugin.toolkit.Pills;
+import com.botmaker.plugin.toolkit.Slots;
 import com.botmaker.sdk.api.geometry.Point;
 import com.botmaker.sdk.api.geometry.Rect;
 import com.botmaker.sdk.api.geometry.Size;